Show synchronization log as a grid in results dialog Improved grid scrolling performance (most noticeable on Linux) Allow grid selection starting from outside of the grid RealTimeSync: Support drag & drop on main dialog for *.ffs_real and *.ffs_batch files Optimized memory consumption when generating log for millions of items Optimized memory consumption when exporting to CSV file Have grid row height match window default font size Catch out of memory when copying huge lists into clipboard Fixed failure to resume aborted sync after having FFS implicitly create target directory Fixed horizontal mouse wheel scrolling direction for RTL languages (Hebrew) RealTimeSync: Fixed drag and drop not working (Linux) Set maximum size of LastSyncs.log in GlobalSettings.xml element <LastSyncsFileSizeMax> Show error when trying to copy a named pipe rather than hang (Linux) Improved copy routine minimizing file accesses (Linux) Copy file access permissions by default (Linux) Fixed unexpected "File or Directory not existing" error during file copy (Linux)

+++ ui/main_dlg.h
@@ -163,7 +163,7 @@ private:
 
     void onGridDoubleClickL(zen::GridClickEvent& event);
     void onGridDoubleClickR(zen::GridClickEvent& event);
-    void onGridDoubleClickRim(int row, bool leftSide);
+    void onGridDoubleClickRim(size_t row, bool leftSide);
 
     void onGridLabelLeftClickC(zen::GridClickEvent& event);
     void onGridLabelLeftClickL(zen::GridClickEvent& event);
